Move decimalToCoin and coinToDecimal into SDK
