+++ packages/bcp/src/decode.ts
@@ -1,4 +1,4 @@
-import { isAminoStdTx, TxsResponse } from "@cosmwasm/sdk";
+import { coinToDecimal, isAminoStdTx, TxsResponse } from "@cosmwasm/sdk";
 import {
   Address,
   Algorithm,
@@ -20,7 +20,7 @@ import {
 import { Encoding } from "@iov/encoding";
 import amino from "@tendermint/amino-js";
 
-import { coinToDecimal, TokenInfos } from "./types";
+import { TokenInfos } from "./types";
 
 const { fromBase64 } = Encoding;
