+++ packages/bcp/src/types.ts
@@ -1,34 +1,7 @@
 import { TokenInfo } from "@cosmwasm/sdk";
 import { Nonce } from "@iov/bcp";
-import { Decimal } from "@iov/encoding";
-import amino from "@tendermint/amino-js";
 
 export type TokenInfos = ReadonlyArray<TokenInfo>;
-
-export function decimalToCoin(lookup: readonly TokenInfo[], value: Decimal, ticker: string): amino.Coin {
-  const match = lookup.find(token => token.ticker === ticker);
-  if (!match) {
-    throw Error(`unknown ticker: ${ticker}`);
-  }
-  if (match.fractionalDigits !== value.fractionalDigits) {
-    throw new Error(
-      "Mismatch in fractional digits between token and value. If you really want, implement a conversion here. However, this indicates a bug in the caller code.",
-    );
-  }
-  return {
-    denom: match.denom,
-    amount: value.atomics,
-  };
-}
-
-export function coinToDecimal(tokens: readonly TokenInfo[], coin: amino.Coin): readonly [Decimal, string] {
-  const match = tokens.find(({ denom }) => denom === coin.denom);
-  if (!match) {
-    throw Error(`unknown denom: ${coin.denom}`);
-  }
-  const value = Decimal.fromAtomics(coin.amount, match.fractionalDigits);
-  return [value, match.ticker];
-}
 
 // tslint:disable-next-line:no-bitwise
 const maxAcct = 1 << 23;
