+++ packages/sdk/package.json
@@ -38,6 +38,7 @@
     "pack-web": "yarn build-or-skip && webpack --mode development --config webpack.web.config.js"
   },
   "dependencies": {
+    "@iov/encoding": "^2.0.0-alpha.7",
     "@tendermint/amino-js": "^0.7.0-alpha.1",
     "axios": "^0.19.0"
   },
