+++ packages/sdk/types/index.d.ts
@@ -1,2 +1,4 @@
+export { coinToDecimal } from "./decoding";
+export { decimalToCoin } from "./encoding";
 export { RestClient, TxsResponse } from "./restclient";
 export { AminoTx, isAminoStdTx, TokenInfo } from "./types";
